+++ src/podofo/main/PdfAcroForm.cpp
@@ -150,21 +150,25 @@ unsigned PdfAcroForm::GetFieldCount() const
 
 PdfAcroForm::iterator PdfAcroForm::begin()
 {
+    initFields();
     return iterator(m_Fields.begin());
 }
 
 PdfAcroForm::iterator PdfAcroForm::end()
 {
+    initFields();
     return iterator(m_Fields.end());
 }
 
 PdfAcroForm::const_iterator PdfAcroForm::begin() const
 {
+    const_cast<PdfAcroForm&>(*this).initFields();
     return const_iterator(m_Fields.begin());
 }
 
 PdfAcroForm::const_iterator PdfAcroForm::end() const
 {
+    const_cast<PdfAcroForm&>(*this).initFields();
     return const_iterator(m_Fields.end());
 }