+++ src/podofo/main/PdfAnnotationCollection.h
@@ -75,6 +75,12 @@ namespace PoDoFo
                 m_iterator++;
                 return *this;
             }
+            Iterator operator++(int)
+            {
+                auto copy = *this;
+                m_iterator++;
+                return copy;
+            }
             value_type operator*()
             {
                 return m_iterator->get();