+++ src/podofo/main/PdfArray.h
@@ -30,7 +30,7 @@ private:
     PdfArrayIndirectIterableBase(PdfArray& arr);
 
 public:
-    class iterator final
+    class Iterator final
     {
         friend class PdfArrayIndirectIterableBase;
     public:
@@ -40,15 +40,16 @@ public:
         using reference = void;
         using iterator_category = std::forward_iterator_tag;
     public:
-        iterator();
+        Iterator();
     private:
-        iterator(TListIterator&& iterator, PdfIndirectObjectList* objects);
+        Iterator(TListIterator&& iterator, PdfIndirectObjectList* objects);
     public:
-        iterator(const iterator&) = default;
-        iterator& operator=(const iterator&) = default;
-        bool operator==(const iterator& rhs) const;
-        bool operator!=(const iterator& rhs) const;
-        iterator& operator++();
+        Iterator(const Iterator&) = default;
+        Iterator& operator=(const Iterator&) = default;
+        bool operator==(const Iterator& rhs) const;
+        bool operator!=(const Iterator& rhs) const;
+        Iterator& operator++();
+        Iterator operator++(int);
         value_type operator*();
         value_type operator->();
     private:
@@ -59,8 +60,8 @@ public:
     };
 
 public:
-    iterator begin() const;
-    iterator end() const;
+    Iterator begin() const;
+    Iterator end() const;
 
 private:
     PdfArray* m_arr;
@@ -401,63 +402,71 @@ PdfArrayIndirectIterableBase<TObject, TListIterator>::PdfArrayIndirectIterableBa
     : PdfIndirectIterableBase(arr), m_arr(&arr) { }
 
 template<typename TObject, typename TListIterator>
-typename PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator PdfArrayIndirectIterableBase<TObject, TListIterator>::begin() const
+typename PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator PdfArrayIndirectIterableBase<TObject, TListIterator>::begin() const
 {
     if (m_arr == nullptr)
-        return iterator();
+        return Iterator();
     else
-        return iterator(m_arr->begin(), GetObjects());
+        return Iterator(m_arr->begin(), GetObjects());
 }
 
 template<typename TObject, typename TListIterator>
-typename PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator PdfArrayIndirectIterableBase<TObject, TListIterator>::end() const
+typename PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator PdfArrayIndirectIterableBase<TObject, TListIterator>::end() const
 {
     if (m_arr == nullptr)
-        return iterator();
+        return Iterator();
     else
-        return iterator(m_arr->end(), GetObjects());
+        return Iterator(m_arr->end(), GetObjects());
 }
 
 template<typename TObject, typename TListIterator>
-PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator::iterator() : m_objects(nullptr) { }
+PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::Iterator() : m_objects(nullptr) { }
 
 template<typename TObject, typename TListIterator>
-PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator::iterator(TListIterator&& iterator, PdfIndirectObjectList* objects)
+PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::Iterator(TListIterator&& iterator, PdfIndirectObjectList* objects)
     : m_iterator(std::move(iterator)), m_objects(objects) { }
 
 template<typename TObject, typename TListIterator>
-bool PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator::operator==(const iterator& rhs) const
+bool PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::operator==(const Iterator& rhs) const
 {
     return m_iterator == rhs.m_iterator;
 }
 
 template<typename TObject, typename TListIterator>
-bool PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator::operator!=(const iterator& rhs) const
+bool PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::operator!=(const Iterator& rhs) const
 {
     return m_iterator != rhs.m_iterator;
 }
 
 template<typename TObject, typename TListIterator>
-typename PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator& PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator::operator++()
+typename PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator& PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::operator++()
 {
     m_iterator++;
     return *this;
 }
 
 template<typename TObject, typename TListIterator>
-typename PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator::value_type PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator::operator*()
+typename PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::operator++(int)
+{
+    auto copy = *this;
+    m_iterator++;
+    return copy;
+}
+
+template<typename TObject, typename TListIterator>
+typename PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::value_type PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::operator*()
 {
     return resolve();
 }
 
 template<typename TObject, typename TListIterator>
-typename PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator::value_type PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator::operator->()
+typename PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::value_type PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::operator->()
 {
     return resolve();
 }
 
 template<typename TObject, typename TListIterator>
-typename PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator::value_type PdfArrayIndirectIterableBase<TObject, TListIterator>::iterator::resolve()
+typename PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::value_type PdfArrayIndirectIterableBase<TObject, TListIterator>::Iterator::resolve()
 {
     TObject& robj = *m_iterator;
     TObject* indirectobj;